+++ Controller/Products/Vers.php
@@ -137,7 +137,9 @@ class Vers extends \Autocompleteplus\Autosuggest\Controller\Products
             'serp_slug' => $this->helper->getSerpSlug($storeId),
             'smart_nav_native' => $this->helper->getSmartNavigationNative($storeId),
             'show_out_of_stock' => $this->helper->getDisplayOutOfStock(),
-            'basic_enabled' => $this->helper->getBasicEnabled($storeId)
+            'basic_enabled' => $this->helper->getBasicEnabled($storeId),
+            'search_engine' => $this->helper->getSearchEngine($storeId),
+            'flat_products_enabled' => $this->helper->getFlatCatalogUsage($storeId)
         ];
 
         $result = $this->resultJsonFactory->create();